+++ Source/Engine/Graphics/Vertex.cpp
@@ -8,8 +8,10 @@ Vertex_P_N_UV_T_B::Vertex_P_N_UV_T_B()
 
 void Vertex_P_N_UV_T_B::AddBone(i32 id, f32 weight)
 {
-  i32 i = std::count_if(std::begin(boneIds), std::end(boneIds), [](u32 id) { return id != -1; });
-  if (i < MAX_BONES_INFLUENCE)
+  i32 i = std::count_if(std::begin(boneIds), std::end(boneIds), [](u32 id)
+                        { return id != static_cast<u32>(-1); });
+
+  if (i < static_cast<i32>(MAX_BONES_INFLUENCE))
   {
     boneIds[i] = id;
     boneWeights[i] = weight;
